+++ lib/openstudio/helpers/string.rb
@@ -3,14 +3,14 @@
 class String
   def underscore
     gsub(/::/, '/')
-      .gsub(/([A-Z]+)([A-Z][a-z])/, '\1_\2')
-      .gsub(/([a-z\d])([A-Z])/, '\1_\2')
-      .tr('-', '_')
-      .downcase
+        .gsub(/([A-Z]+)([A-Z][a-z])/, '\1_\2')
+        .gsub(/([a-z\d])([A-Z])/, '\1_\2')
+        .tr('-', '_')
+        .downcase
   end
 
   def snake_case
-    gsub(' ', '_').downcase
+    return gsub(/\s/, '_').squeeze('_').downcase
   end
 
   def to_bool