+++ spec/files/measures/rotate_building/measure.rb
@@ -1,0 +1,5 @@
+class RotateBuilding < OpenStudio::Ruleset::ModelUserScript
+  # not a real measure
+end # end the measure
+
+RotateBuilding.new.registerWithApplication