+++ spec/openstudio/formulation_spec.rb
@@ -48,6 +48,8 @@ describe OpenStudio::Analysis::Formulation do
       objective_function: true
                  )
 
+    puts a.outputs
+    
     expect(a.to_hash[:analysis][:output_variables].first[:objective_function_index]).to eq 0
 
     a.add_output(